services/ml-service/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Machine Learning
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.1

# MLOps
mlflow==2.7.1
bentoml==1.1.6

# Model explainability
shap==0.42.1

# Data processing
pydantic==2.4.2
pydantic-settings==2.0.3

# Database and caching
redis==5.0.1
asyncpg==0.28.0
sqlalchemy==2.0.21

# Environment and configuration
python-dotenv==1.0.0

# Monitoring and logging
prometheus-client==0.17.1
structlog==23.1.0

# Development
pytest==7.4.2
pytest-asyncio==0.21.1
black==23.9.1
flake8==6.1.0
mypy==1.6.1